+++ app/src/main/java/org/simple/clinic/analytics/UpdateAnalyticsUserId.kt
@@ -7,6 +7,7 @@ import org.simple.clinic.user.User.LoggedInStatus.LOGGED_IN
 import org.simple.clinic.user.User.LoggedInStatus.RESETTING_PIN
 import org.simple.clinic.user.User.LoggedInStatus.RESET_PIN_REQUESTED
 import org.simple.clinic.user.UserSession
+import org.simple.clinic.util.extractIfPresent
 import org.simple.clinic.util.filterAndUnwrapJust
 import org.simple.clinic.util.scheduler.SchedulersProvider
 import javax.inject.Inject
@@ -24,8 +25,7 @@ class UpdateAnalyticsUserId @Inject constructor(
   fun listen() {
     userSession
         .loggedInUser()
-        .take(1)
-        .filterAndUnwrapJust()
+        .extractIfPresent()
         .filter { it.loggedInStatus in statesToSetUserIdFor }
         .map { AnalyticsUser(it.uuid) }
         .doOnNext(Analytics::setLoggedInUser)