+++ app/src/main/java/org/simple/clinic/plumbing/infrastructure/DatadogInfrastructure.kt
@@ -18,4 +18,8 @@ class DatadogInfrastructure @Inject constructor() : Infrastructure {
         )
     )
   }
+
+  override fun clear() {
+    Datadog.setUserInfo(id = null, extraInfo = emptyMap())
+  }
 }