+++ app/src/main/java/org/simple/clinic/plumbing/infrastructure/Infrastructure.kt
@@ -11,4 +11,6 @@ interface Infrastructure {
       country: Country,
       deployment: Deployment
   )
+
+  fun clear()
 }