+++ app/src/main/java/org/simple/clinic/plumbing/infrastructure/SentryInfrastructure.kt
@@ -6,12 +6,26 @@ import org.simple.clinic.appconfig.Deployment
 import org.simple.clinic.user.User
 import javax.inject.Inject
 
-class SentryInfrastructure @Inject constructor(): Infrastructure {
+class SentryInfrastructure @Inject constructor() : Infrastructure {
+
+  companion object {
+    private const val TAG_USER_ID = "userUuid"
+    private const val TAG_FACILITY_ID = "facilityUuid"
+    private const val TAG_COUNTRY_CODE = "countryCode"
+    private const val TAG_DEPLOYMENT = "deployment"
+  }
 
   override fun addDetails(user: User, country: Country, deployment: Deployment) {
-    Sentry.setTag("userUuid", user.uuid.toString())
-    Sentry.setTag("facilityUuid", user.currentFacilityUuid.toString())
-    Sentry.setTag("countryCode", country.isoCountryCode)
-    Sentry.setTag("deployment", deployment.endPoint.toString())
+    Sentry.setTag(TAG_USER_ID, user.uuid.toString())
+    Sentry.setTag(TAG_FACILITY_ID, user.currentFacilityUuid.toString())
+    Sentry.setTag(TAG_COUNTRY_CODE, country.isoCountryCode)
+    Sentry.setTag(TAG_DEPLOYMENT, deployment.endPoint.toString())
+  }
+
+  override fun clear() {
+    Sentry.removeTag(TAG_USER_ID)
+    Sentry.removeTag(TAG_FACILITY_ID)
+    Sentry.removeTag(TAG_COUNTRY_CODE)
+    Sentry.removeTag(TAG_DEPLOYMENT)
   }
 }