+++ app/src/main/java/org/simple/clinic/settings/SettingsEffect.kt
@@ -12,4 +12,10 @@ object CheckAppUpdateAvailable : SettingsEffect()
 
 sealed class SettingsViewEffect : SettingsEffect()
 
+object LogoutUser : SettingsEffect()
+
 object OpenLanguageSelectionScreenEffect : SettingsViewEffect()
+
+object ShowConfirmLogoutDialog : SettingsViewEffect()
+
+object RestartApp : SettingsViewEffect()