+++ app/src/main/java/org/simple/clinic/settings/SettingsViewEffectHandler.kt
@@ -10,6 +10,8 @@ class SettingsViewEffectHandler(
   override fun handle(viewEffect: SettingsViewEffect) {
     when (viewEffect) {
       OpenLanguageSelectionScreenEffect -> uiActions.openLanguageSelectionScreen()
+      ShowConfirmLogoutDialog -> uiActions.showConfirmLogoutDialog()
+      RestartApp -> uiActions.restartApp()
     }.exhaustive()
   }
 }