+++ app/src/main/java/org/simple/clinic/settings/UiActions.kt
@@ -3,4 +3,6 @@ package org.simple.clinic.settings
 interface UiActions {
 
   fun openLanguageSelectionScreen()
+  fun showConfirmLogoutDialog()
+  fun restartApp()
 }